refac: Move storage column name classes to better locations (#2620)
