+++ source/databricks/calculation_engine/tests/databases/wholesale_results_internal/test_monthly_amounts_per_charge_storage_model_factory.py
@@ -32,8 +32,8 @@ from package.codelists import (
     ChargeUnit,
     CalculationType,
 )
-from package.constants import (
-    Colname,
+from package.constants import Colname
+from package.databases.wholesale_results_internal.monthly_amounts_column_names import (
     MonthlyAmountsColumnNames,
 )
 from package.infrastructure.paths import (
